Mundo 1 python

+++ exe01.py
@@ -1,6 +1,6 @@
 
-num1=float(input('digite o primeiro numero:'))
-num2=float(input('digite o segundo numero:'))
+num1=int(input('digite o primeiro numero:'))
+num2=int(input('digite o segundo numero:'))
 
 
 S=num1+num2
